Rename a few files, add specific test for go18
